add cmake, llvm and cmake

+++ nbre/prepare.sh
@@ -19,14 +19,61 @@
 
 #CUR_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}"  )" >/dev/null && pwd  )"
 CUR_DIR="$( pwd )"
+
+if [ ! -d $CUR_DIR/3rd_party/cmake-3.12.2 ]; then
+  tar -xf cmake-3.12.2.tar.gz
+fi
+
+if [ ! -f $CUR_DIR/lib/bin/cmake ]; then
+  cd $CUR_DIR/3rd_party/cmake-3.12.2/
+  ./bootstrap --prefix=$CUR_DIR/lib --parallel=4 && make && make install
+fi
+export PATH=$CUR_DIR/lib/bin:$PATH
+
+cd $CUR_DIR/3rd_party
+LLVM_VERSION=6.0.1
+unzip_llvm_tar(){
+  if [ ! -d $1-$LLVM_VERSION.src ]; then
+    tar -xf $1-$LLVM_VERSION.src.tar.xz
+  fi
+}
+unzip_llvm_tar llvm
+unzip_llvm_tar cfe
+unzip_llvm_tar clang-tools-extra
+unzip_llvm_tar compiler-rt
+unzip_llvm_tar libcxx
+unzip_llvm_tar libcxxabi
+unzip_llvm_tar libunwind
+unzip_llvm_tar lld
+
+if [ ! -d $CUR_DIR/lib/include/llvm ]; then
+  ln -s $CUR_DIR/3rd_party/cfe-$LLVM_VERSION.src $CUR_DIR/3rd_party/llvm-$LLVM_VERSION.src/tools/clang
+  ln -s $CUR_DIR/3rd_party/lld-$LLVM_VERSION.src $CUR_DIR/3rd_party/llvm-$LLVM_VERSION.src/tools/lld
+  ln -s $CUR_DIR/3rd_party/clang-tools-extra-$LLVM_VERSION.src $CUR_DIR/3rd_party/llvm-$LLVM_VERSION.src/tools/clang/tools/extra
+  ln -s $CUR_DIR/3rd_party/compiler-rt-$LLVM_VERSION.src $CUR_DIR/3rd_party/llvm-$LLVM_VERSION.src/projects/compiler-rt
+  ln -s $CUR_DIR/3rd_party/libcxx-$LLVM_VERSION.src $CUR_DIR/3rd_party/llvm-$LLVM_VERSION.src/projects/libcxx
+  ln -s $CUR_DIR/3rd_party/libcxxabi-$LLVM_VERSION.src $CUR_DIR/3rd_party/llvm-$LLVM_VERSION.src/projects/libcxxabi
+
+  cd $CUR_DIR/3rd_party
+  mkdir llvm-build
+  cd llvm-build
+  cmake -DCMAKE_BUILD_TYPE=Release -DCMAKE_INSTALL_PREFIX=$CUR_DIR/lib/ ../llvm-$LLVM_VERSION.src
+  make -j4 && make install
+fi
+
+export CXX=$CUR_DIR/lib/bin/clang++
+export CC=$CUR_DIR/lib/bin/clang
+
 cd $CUR_DIR/3rd_party
 if [ ! -d "boost_1_67_0"  ]; then
   tar -zxvf boost_1_67_0.tar.gz
 fi
-cd boost_1_67_0
-./bootstrap.sh --prefix=$CUR_DIR/lib/ 
-./b2 --toolset=clang
-./b2 install
+if [ ! -d $CUR_DIR/lib/include/boost ]; then
+  cd boost_1_67_0
+  ./bootstrap.sh --prefix=$CUR_DIR/lib/
+  ./b2 --toolset=clang
+  ./b2 install
+fi
 
 build_with_cmake(){
   cd $CUR_DIR/3rd_party/$1
@@ -35,7 +82,7 @@ build_with_cmake(){
   fi
   mkdir build
   cd build
-  cmake -DCMAKE_LIBRARY_PATH=$CUR_DIR/lib/lib -DCPATH=$CUR_DIR/lib/include -DCMAKE_BUILD_TYPE=Release -DCMAKE_INSTALL_PREFIX=$CUR_DIR/lib/ ../
+  cmake -DCMAKE_MODULE_PATH=$CUR_DIR/lib/lib/cmake -DCMAKE_LIBRARY_PATH=$CUR_DIR/lib/lib -DCMAKE_BUILD_TYPE=Release -DCMAKE_INSTALL_PREFIX=$CUR_DIR/lib/ ../
   make && make install && make clean
   cd ../ && rm -rf build
 }
@@ -51,24 +98,43 @@ build_with_make(){
   make && make install PREFIX=$CUR_DIR/lib/
 }
 
-build_with_cmake glog
-build_with_cmake gtest
-cd $CUR_DIR/3rd_patch/snappy && cp ../snappy.patch ./ && git apply snappy.patch
-build_with_cmake snappy
+if [ ! -d $CUR_DIR/lib/include/glog/ ]; then
+  build_with_cmake glog
+fi
+if [ ! -d $CUR_DIR/lib/include/gtest/ ]; then
+  build_with_cmake googletest
+fi
 
-build_with_configure gflags
-build_with_configure zlib
+if [ ! -f $CUR_DIR/lib/include/snappy.h ]; then
+  cd $CUR_DIR/3rd_patch/snappy && cp ../snappy.patch ./ && git apply snappy.patch
+  build_with_cmake snappy
+fi
+
+if [ ! -d $CUR_DIR/lib/include/gflags ]; then
+  build_with_configure gflags
+fi
+
+if [ ! -f $CUR_DIR/lib/include/zlib.h ]; then
+  build_with_configure zlib
+fi
 
 cd $CUR_DIR/3rd_party
 if [ ! -d "zstd-1.1.3"  ]; then
   tar -zxvf zstd-1.1.3.tar.gz
 fi
-build_with_make zstd-1.1.3
+if [ ! -f $CUR_DIR/lib/include/zstd.h ]; then
+  build_with_make zstd-1.1.3
+fi
+if [ ! -f $CUR_DIR/lib/include/bzlib.h ]; then
 build_with_make bzip2-1.0.6
-build_with_make lz4
+fi
+if [ ! -f $CUR_DIR/lib/include/lz4.h ]; then
+  build_with_make lz4
+fi
 
-cd $CUR_DIR/3rd_party/rocksdb
-LIBRARY_PATH=$CUR_DIR/lib/lib CPATH=$CUR_DIR/lib/include make install-static INSTALL_PATH=$CUR_DIR/lib -j4
-
+if [ ! -d $CUR_DIR/lib/include/rocksdb ]; then
+  cd $CUR_DIR/3rd_party/rocksdb
+  LIBRARY_PATH=$CUR_DIR/lib/lib CPATH=$CUR_DIR/lib/include make install-static INSTALL_PATH=$CUR_DIR/lib -j4
+fi
